backend/models/vehicle_counter.py
"""
Vehicle Counting Model
Uses YOLO/ML model to count vehicles in traffic images
"""
import random
from PIL import Image

# TODO: Replace with actual YOLO model
# from ultralytics import YOLO
# from backend.config import config
# model = YOLO(config.YOLO_MODEL)

def count_vehicles(image: Image.Image) -> int:
    """
    Count vehicles in the given image
    
    Args:
        image: PIL Image object
        
    Returns:
        Number of vehicles detected
    """
    # TODO: Implement actual YOLO detection
    # Currently using stub for demo
    
    # results = model(image)
    # vehicle_count = len(results[0].boxes)
    
    # Stub implementation
    vehicle_count = random.randint(8, 145)
    
    return vehicle_count


def get_vehicle_details(image: Image.Image) -> dict:
    """
    Get detailed vehicle information
    
    Args:
        image: PIL Image object
        
    Returns:
        Dictionary with vehicle type breakdown
    """
    # TODO: Implement actual vehicle classification
    
    total = count_vehicles(image)
    
    # Stub breakdown
    return {
        "total": total,
        "cars": int(total * 0.6),
        "bikes": int(total * 0.25),
        "trucks": int(total * 0.10),
        "buses": int(total * 0.05)
    }
